refactor static generator
