--- src/templates/my-post/blog-post.module.css.d.ts
@@ -1,6 +1,0 @@
-// This file is automatically generated. Do not modify this file manually -- YOUR CHANGES WILL BE ERASED!
-export const meta: string;
-export const article: string;
-export const body: string;
-export const articleNavigation: string;
-export const container: string;
